[corrections last merge] minors corrections following the precedent merge
	modified:   includes/init.h
	modified:   includes/shared.h
	modified:   includes/tools.h
	modified:   src/init/3_mapcontent/init_3_mapcontent_main.c
	modified:   src/init/3_mapcontent/init_3_mapcontent_utils.c
	modified:   src/tools/tools_time.c

+++ includes/tools.h
@@ -6,7 +6,7 @@
 /*   By: acardona <acardona@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/19 14:10:26 by acardona          #+#    #+#             */
-/*   Updated: 2023/10/01 00:11:44 by acardona         ###   ########.fr       */
+/*   Updated: 2023/10/01 00:17:19 by acardona         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,7 @@
 # include <X11/X.h>
 # include <X11/keysym.h>
 # include <stdio.h>
+# include <sys/time.h>
 
 # include "../mlx/mlx.h"
 # include "../libft/libft.h"
@@ -54,17 +55,17 @@ typedef struct s_coord_i
 	int	y;
 }	t_coord_i;
 
-void			to_vector_print(t_vector_f v);
-double			to_vector_norm(t_vector_f v0, t_vector_f v1);
+void	to_vector_print(t_vector_f v);
+double	to_vector_norm(t_vector_f v0, t_vector_f v1);
 
 //tools_error
-void			to_error_msg(char *msg);
+void	to_error_msg(char *msg);
 
 //tools_lst.c
-void			to_lstfree(t_list **lst);
+void	to_lstfree(t_list **lst);
 
 //tools_time.c
-unsigned int	to_getime(void);
+size_t	to_getime(void);
 
 //tools_anglesset.c
 void	to_angle_set_init(float *old_fov, float new_fov,
